V 0.9.2.0
- Addition of "SimpleIC - ImplementBalls"

With this addition to the script it is possible to add those balls for the lower arm attachers to implements that don't have them. It is completely global, inserts itself to each implement that has an implement attacher.
No need to add to the individual implements.

Its as simple as walking up to the implement attacher and hitting the X with the mouse as usual with simpleIC.